backend/requirements/base.txt
Django~=3.0
djangorestframework~=3.11
django-cors-headers~=3.2
django-filter~=2.2
gunicorn~=20.0
django-hosts~=4.0
Pygments~=2.6
markdown~=3.2
readtime~=1.1
martor~=1.4
Pillow~=7.1
django-cleanup~=4.0
requests~=2.23
django-import-export~=2.0